+++ doc/changelog.md
@@ -2,6 +2,20 @@
 
 # Change Log
 
+## May-12-2023: Version 2.19.1
+
+- Improvement of parsing forces of the built-in LAMMPS interface.
+- Update of the ABACUS interface.
+- Changed `--symmetry` option to write also `phonopy_symells.yaml`.
+
+## May-12-2023: Version 2.19.0
+
+- Experimental release of built-in LAMMPS interface.
+
+## Mar-30-2023: Version 2.18.0
+
+- Maintenance release. Refactoring of `PhonopyYaml` class.
+
 ## Mar-6-2023: Version 2.17.2
 
 - Bug fixes of user interface.